backend/requirements.txt
Flask==2.3.2
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
firebase-admin==6.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
werkzeug==2.3.7